cleaned up code and fixed bug

+++ server/src/test/java/server/controller/AppControllerTest.java
@@ -54,13 +54,13 @@ class AppControllerTest {
     void updateLobbyPlayerTime() {
         assertEquals(p3, ctrl.updateLobbyPlayerTime("Kelly"));
         long millisecondDif = new Date().getTime() - p3.getTimestamp().getTime();
-        assertTrue(5000L > millisecondDif);
+        assertTrue(5000 > (int) millisecondDif);
     }
 
     @Test
     void updateGamePlayerTime() {
         assertEquals(p1, ctrl.updateGamePlayerTime(game.getId(), "Charlie"));
         long millisecondDif = new Date().getTime() - p3.getTimestamp().getTime();
-        assertTrue(5000L > millisecondDif);
+        assertTrue(5000 > (int) millisecondDif);
     }
 }
